+++ tests/testthat/test-mighty_study.R
@@ -141,3 +141,13 @@ test_that("validate_datasets() error on incorrect file name", {
   files <- c("example/adae.yaml", "example/advs.yaml", "example/_test.yaml")
   expect_error(validate_datasets(files))
 })
+
+test_that("validate_datasets() accepts AD* dataset file names", {
+  files <- c("example/adae.yaml", "example/advs.yaml")
+  expect_no_error(validate_datasets(files))
+})
+
+test_that("validate_datasets() accepts MD* dataset file names", {
+  files <- c("example/mdcol.yaml", "example/mdparam.yaml")
+  expect_no_error(validate_datasets(files))
+})